lorry: svn add branches and tags after clone
Glob expressions can be used to specify branches, but git svn clone/init will append /* to
them unless they have a * in them. It is possible to have globs without.
netpbm has branches in its root directory, so the glob
{advanced,stable,super-stable} should select only those directories
as branches, but they become {a,s,ss}/*, so subdirectories of each become
branches

The workaround is to manually add the glob expression to config
so it is cloned then branches are added, then re fetched