lorry: store the marks in the .git directory
The marks correspond to the import state of the git repository,
the .git directory is where the current state of the git repo is kept,
so it makes sense for the marks to be in there.

This is partly to allow the git state to be backed up before
a repository is processed in case it is left in an inconsistent state.