lorries: delete some, move others
This should preferrably be the active set of lorries.
If one you need has been moved or removed, please re-add it.

cairo, docbook, drm, fontutils, freetype, libjpeg, libpipeline,
libpng, libtiff, mesa, pixman, shared-mime-info and xterm are
probably needed for X, but strata for those are not in master.

less is available in busybox, but the busybox less is pretty poor,
so it may be wanted later.

wget and rsync were once needed to build, this should not be the case
so they have been moved

lsh and gnutls have been moved since they were needed for wget, which
is not built any more, however something will need gnutls, so keep it
elsewhere.

libtirpc, portmap, rpcbind and nfs-utils are needed for proper nfs
support, busybox works without locking for now, but we will want a
proper implementation at some point

udev is in systemd now, so it has been removed
bc was once used in the bootstrap script, this is no longer the case
and it should not be brought back, so removed
binutils is available under the name binutils-redhat, which is already
converted to git
fakeroot, kbd, parted and squashfs-tools were never used
inetutils, procps and psmisc are suitably replaced by busybox
sudo was never properly set up and the typical use for us is not
multi-user systems so it isn't needed
gnulib is handled by submodule support