lorry: push to url specified in lorry config
Rather than pushing to the remote 'gitorious', it is better to push
to the remote specified in lorry's config, since it may have changed
since the repository was originally mirrored, so the remote in
git's config may be out of date.

This may make the setting of the 'gitirous' remote obsolete now.

Note that the semantic difference between pushing to a remote and
pushing to a url isn't triggered because we always pass refspecs.